pulled logic out of MainActivity into LogicSubsystem
