+++ app/src/main/java/com/evanv/taskapp/ui/main/MainActivity.java
@@ -1,15 +1,12 @@
 package com.evanv.taskapp.ui.main;
 
-import static com.evanv.taskapp.logic.Event.clearTime;
 import static com.evanv.taskapp.logic.Task.clearDate;
-import static com.evanv.taskapp.logic.Task.getDiff;
 
 import android.content.DialogInterface;
 import android.content.Intent;
 import android.content.SharedPreferences;
 import android.os.Bundle;
 import android.text.InputType;
-import android.util.Log;
 import android.view.Menu;
 import android.view.MenuItem;
 import android.widget.EditText;
@@ -23,39 +20,23 @@ import androidx.annotation.Nullable;
 import androidx.appcompat.app.AlertDialog;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.coordinatorlayout.widget.CoordinatorLayout;
-import androidx.lifecycle.ViewModelProvider;
-import androidx.lifecycle.ViewModelStoreOwner;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 
 import com.evanv.taskapp.R;
 import com.evanv.taskapp.databinding.ActivityMainBinding;
-import com.evanv.taskapp.db.TaskAppViewModel;
-import com.evanv.taskapp.logic.Event;
-import com.evanv.taskapp.logic.Optimizer;
-import com.evanv.taskapp.logic.Task;
+import com.evanv.taskapp.logic.LogicSubsystem;
 import com.evanv.taskapp.ui.additem.AddItem;
-import com.evanv.taskapp.ui.additem.recur.DailyRecurFragment;
-import com.evanv.taskapp.ui.additem.recur.MonthlyRecurFragment;
-import com.evanv.taskapp.ui.additem.recur.NoRecurFragment;
-import com.evanv.taskapp.ui.additem.recur.RecurActivity;
-import com.evanv.taskapp.ui.additem.recur.RecurInput;
-import com.evanv.taskapp.ui.additem.recur.WeeklyRecurFragment;
-import com.evanv.taskapp.ui.additem.recur.YearlyRecurFragment;
 import com.evanv.taskapp.ui.main.recycler.DayItem;
 import com.evanv.taskapp.ui.main.recycler.DayItemAdapter;
-import com.evanv.taskapp.ui.main.recycler.EventItem;
-import com.evanv.taskapp.ui.main.recycler.TaskItem;
 import com.google.android.material.behavior.HideBottomViewOnScrollBehavior;
 import com.google.android.material.floatingactionbutton.FloatingActionButton;
 
-import java.text.ParseException;
 import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.Calendar;
 import java.util.Date;
 import java.util.List;
-import java.util.Objects;
+
+import kotlin.Pair;
 
 /**
  * Main Activity for the app. Display's the user's schedule of Tasks/Events, while allowing for
@@ -68,20 +49,14 @@ public class MainActivity extends AppCompatActivity implements ClickListener {
     // Fields
     @SuppressWarnings("unused")
     private ActivityMainBinding mBinding;          // Binding for the MainActivity
-    private Date mStartDate;                       // The current date
-    private int mTodayTime;                        // The time spent completing tasks today
     public static final int ITEM_REQUEST = 1;      // requestCode for task/item entry
     private DayItemAdapter mDayItemAdapter;        // Adapter for recycler showing user commitments
-    private List<Task> mTasks = new ArrayList<>(); // List of all tasks for user
-    private TaskAppViewModel mTaskAppViewModel;    // ViewModel to interact with Database
     private ViewFlipper mVF;                       // Swaps between loading screen and recycler
-    // taskSchedule[i] represents the list of tasks for the day i days past startDate
-    private final ArrayList<ArrayList<Task>> mTaskSchedule = new ArrayList<>();
-    // eventSchedule[i] represents the list of events for the day i days past startDate
-    private final ArrayList<ArrayList<Event>> mEventSchedule = new ArrayList<>();
     // Allows data to be pulled from activity
     private ActivityResultLauncher<Intent> mStartForResult;
     // Allows us to manually show FAB when task/event completed/deleted.
+    LogicSubsystem mLogicSubsystem;                // Subsystem that handles logic for taskapp
+    private Date mStartDate;                       // The current date
 
     // Key for the extra that stores the list of Task names for the Parent Task Picker Dialog in
     public static final String EXTRA_TASKS = "com.evanv.taskapp.ui.main.extras.TASKS";
@@ -89,67 +64,6 @@ public class MainActivity extends AppCompatActivity implements ClickListener {
     private static final String PREF_FILE = "taskappPrefs"; // File name for sharedPrefs
     private static final String PREF_DAY = "taskappDay";    // Day for todayTime
     private static final String PREF_TIME = "taskappTime";  // Time for todayTime
-
-    /**
-     * Removes a task from the task dependency graph, while asking the user how long it took to
-     * complete and adding that time to todayTime to prevent overscheduling on today's date.
-     *
-     * @param task The task to be removed from the task dependency graph
-     * @param showDialog true if dialog is needed, false if dialog isn't
-     * @param refresh true if we should refresh recycler
-     */
-    private void Complete(Task task, boolean showDialog, boolean refresh) {
-        // Show loading screen
-        mTasks.remove(task);
-
-        Date doDate = task.getDoDate();
-
-        // Get the number of days past the start date this task is scheduled for, so we can get the
-        // index of the taskSchedule member for it's do date.
-        int diff = getDiff(doDate, mStartDate);
-
-        // If the task is in the internal data structure, remove it.
-        if (diff >= 0) {
-            mTaskSchedule.get(diff).remove(task);
-            mDayItemAdapter.mDayItemList.set(diff, DayItemHelper(diff));
-            mDayItemAdapter.notifyItemChanged(diff);
-        }
-
-        // Remove the task from the task dependency graph
-        for (int i = 0; i < task.getChildren().size(); i++) {
-            task.getChildren().get(i).removeParent(task);
-        }
-        for (int i = 0; i < task.getParents().size(); i++) {
-            task.getParents().get(i).removeChild(task);
-        }
-
-        if (showDialog) {
-            // Prompt the user to ask how long it took to complete the task, and add this time to
-            // todayTime to prevent the user from being overscheduled on today's date.
-            AlertDialog.Builder builder = new AlertDialog.Builder(this);
-            builder.setMessage(String.format(getString(R.string.complete_dialog_message),
-                    task.getName()));
-            builder.setTitle(R.string.complete_dialog_title);
-
-            final EditText input = new EditText(this);
-            input.setInputType(InputType.TYPE_CLASS_NUMBER);
-            builder.setView(input);
-
-            builder.setPositiveButton(R.string.complete_task, (dialogInterface, i) -> {
-                mTodayTime += Integer.parseInt(input.getText().toString());
-                // As the task dependency graph has been updated, we must reoptimize it
-                Optimize(refresh);
-                // Show recycler
-                mVF.setDisplayedChild(1);
-            });
-
-            builder.show();
-        }
-        else {
-            // As the task dependency graph has been updated, we must reoptimize it
-            Optimize(refresh);
-        }
-    }
 
     /**
      * Handles activities started for a result, in this case when the AddItem activity returns with
@@ -165,7 +79,7 @@ public class MainActivity extends AppCompatActivity implements ClickListener {
         mVF.setDisplayedChild(0);
 
         if (resultCode != RESULT_OK) {
-            if (mTaskSchedule.size() != 0 || mEventSchedule.size() != 0) {
+            if (!mLogicSubsystem.isEmpty()) {
                 mVF.setDisplayedChild(1);
             }
             else {
@@ -173,658 +87,20 @@ public class MainActivity extends AppCompatActivity implements ClickListener {
             }
         }
 
-        // Get the data to build the item
-        Bundle result = Objects.requireNonNull(data).getBundleExtra(AddItem.EXTRA_ITEM);
-        String type = result.getString(AddItem.EXTRA_TYPE);
-
-        // If the item type is Event
-        if (type.equals(AddItem.EXTRA_VAL_EVENT)) {
-            // Get the fields from the bundle
-            String name = result.getString(AddItem.EXTRA_NAME);
-            int ttc = Integer.parseInt(result.getString(AddItem.EXTRA_TTC));
-            String startStr = result.getString(AddItem.EXTRA_START);
-            Bundle recur = result.getBundle(AddItem.EXTRA_RECUR);
-
-            // Convert the String start time into a MyTime
-            Date start;
-            Calendar userStart;
-            try {
-                start = Event.dateFormat.parse(startStr);
-                userStart = Calendar.getInstance();
-                if (start != null) {
-                    userStart.setTime(start);
-                }
-                else {
-                    return;
-                }
-            }
-            catch (Exception e) {
-                System.out.println(e.getMessage());
-                return;
-            }
-
-            // Calculate how many days past today's date this event is scheduled (used to
-            // index into eventSchedule
-            int diff = getDiff(start, mStartDate);
-
-            // Add the new event to the data structure
-            if (!recur.get(RecurInput.EXTRA_TYPE).equals(NoRecurFragment.EXTRA_VAL_TYPE)) {
-                boolean until = recur.get(RecurActivity.EXTRA_UNTIL_TYPE)
-                        .equals(RecurActivity.EXTRA_VAL_UNTIL);
-
-                // Initialize until holders. Only the one associated with the user's choice
-                // of recurrence is used.
-                int endIndex = -1;
-                int numTimes = -1;
-                Date endDate = new Date();
-
-                // Recur until end date
-                if (until) {
-                    try {
-                        endDate = Task.dateFormat.parse
-                                (recur.getString(RecurActivity.EXTRA_UNTIL));
-                        endIndex = getDiff(endDate, mStartDate);
-                    } catch (ParseException e) {
-                        Log.e("TaskApp.MainActivity", e.getMessage());
-                    }
-                }
-                // Recur set number of times
-                else {
-                    numTimes = Integer.parseInt(recur.getString(RecurActivity.EXTRA_UNTIL));
-                }
-
-                String recurType = recur.getString(RecurInput.EXTRA_TYPE);
-                if (DailyRecurFragment.EXTRA_VAL_TYPE.equals(recurType)) {
-                    // Number of days between recurrences
-                    int interval = recur.getInt(DailyRecurFragment.EXTRA_INTERVAL);
-
-                    // Calculates the number of times recurrence occurs.
-                    numTimes = (until) ? ((endIndex - diff) / interval) + 1 : numTimes;
-
-                    // Make sure there is enough mEventSchedules.
-                    for (int i = mEventSchedule.size(); i <= diff + (numTimes * interval);
-                         i++) {
-                        mEventSchedule.add(new ArrayList<>());
-                        mDayItemAdapter.notifyItemChanged(i);
-                    }
-
-                    // Add the event once every interval days.
-                    for (int i = 0; i < (numTimes * interval); i += interval) {
-                        // Calculate date to add event on.
-                        Calendar recurCal = Calendar.getInstance();
-                        recurCal.setTime(start);
-                        recurCal.add(Calendar.DAY_OF_YEAR, i);
-
-                        Event toAdd = new Event(name, ttc, recurCal.getTime());
-                        mEventSchedule.get(diff + i).add(toAdd);
-                        mTaskAppViewModel.insert(toAdd);
-                        mDayItemAdapter.notifyItemChanged(diff + i);
-                    }
-                }
-                else if (WeeklyRecurFragment.EXTRA_VAL_TYPE.equals(recurType)) {
-                    // Number of days between recurrences
-                    int interval = recur.getInt(WeeklyRecurFragment.EXTRA_INTERVAL);
-                    // day[0] = if recurs on sunday, day[6] if recurs on saturday
-                    boolean[] days = recur.getBooleanArray(WeeklyRecurFragment.EXTRA_DAYS);
-
-                    for (int i = 1; i <= 7; i++) {
-                        if (days[i - 1]) {
-                            Calendar recurCal = Calendar.getInstance();
-                            recurCal.setTime(start);
-
-                            // Calculates how many days we must add to recurCal to get next
-                            // day of week
-                            int currIndex = ((i - recurCal.get(Calendar.DAY_OF_WEEK)) + 7)
-                                    % 7;
-
-                            // oneLess shows us if this day occurs before the initial day of the
-                            // week. This is useful as it stops the days before from being
-                            // skewed away from the actual repetition
-                            boolean oneLess = (i - recurCal.get(Calendar.DAY_OF_WEEK)) < 0;
-
-                            recurCal.add(Calendar.DAY_OF_YEAR, currIndex);
-
-                            // Gets start index into eventSchedule
-                            currIndex += diff;
-
-                            if (until) {
-                                numTimes = Integer.MAX_VALUE;
-                            }
-                            else {
-                                endDate.setTime(Long.MAX_VALUE);
-                            }
-
-                            // This means that our week iterator is off by one.
-                            boolean offByOne = oneLess && (interval != 1);
-
-                            for (int j = 0; j < numTimes && !recurCal.getTime().after(endDate);
-                                 j++) {
-                                // Add event to schedule
-                                // Make sure an edge case isn't reached. oneLess only matters if
-                                // interval == 1, as the offByOne if clause fixes it otherwise
-                                if (!(j == 0 && offByOne) && !(j == numTimes - 1 && oneLess &&
-                                        !offByOne)) {
-                                    Event toAdd = new Event(name, ttc, recurCal.getTime());
-
-                                    for (int k = mEventSchedule.size(); k <= currIndex; k++) {
-                                        mEventSchedule.add(new ArrayList<>());
-                                        mDayItemAdapter.notifyItemChanged(k);
-                                    }
-
-                                    mEventSchedule.get(currIndex).add(toAdd);
-                                    mTaskAppViewModel.insert(toAdd);
-                                    mDayItemAdapter.notifyItemChanged(currIndex);
-                                }
-
-                                // Calculate next date
-                                recurCal.add(Calendar.WEEK_OF_YEAR, interval);
-                                currIndex += (7 * interval);
-
-                                // Fix the skew caused by an event recurring at a non-1 interval
-                                if (j == 0 && offByOne) {
-                                    recurCal.add(Calendar.WEEK_OF_YEAR, -1);
-                                    currIndex -= 7;
-                                }
-                            }
-                        }
-                    }
-                }
-                // User chose to recur monthly
-                else if (MonthlyRecurFragment.EXTRA_VAL_TYPE.equals(recurType)) {
-                    // Get recur interval
-                    int interval = recur.getInt(MonthlyRecurFragment.EXTRA_INTERVAL);
-                    String intervalType = recur.getString(MonthlyRecurFragment.EXTRA_RECUR_TYPE);
-
-                    Calendar recurCal = Calendar.getInstance();
-                    recurCal.setTime(start);
-
-                    int currIndex = diff;
-
-                    // User chose to recur on the same day every month
-                    if (intervalType.equals(MonthlyRecurFragment.EXTRA_VAL_STATIC)) {
-                        // Calculate when to recur until
-                        if (!until) {
-                            Calendar endDateCal = Calendar.getInstance();
-                            endDateCal.setTime(start);
-                            endDateCal.add(Calendar.MONTH, interval * (numTimes - 1));
-                            endDate = endDateCal.getTime();
-                            endIndex = getDiff(endDate, mStartDate);
-                        }
-
-                        // Ensure there is enough space in the eventSchedule
-                        for (int i = mEventSchedule.size(); i <= endIndex; i++) {
-                            mEventSchedule.add(new ArrayList<>());
-                            mDayItemAdapter.notifyItemChanged(i);
-                        }
-
-                        // Add an event at a consistent interval until end date is reached
-                        while (!recurCal.getTime().after(endDate)) {
-                            // Add event to schedule
-                            Event toAdd = new Event(name, ttc, recurCal.getTime());
-                            mEventSchedule.get(currIndex).add(toAdd);
-                            mTaskAppViewModel.insert(toAdd);
-                            mDayItemAdapter.notifyItemChanged(currIndex);
-
-                            // Calculate next date
-                            recurCal.add(Calendar.MONTH, interval);
-                            currIndex = getDiff(recurCal.getTime(), mStartDate);
-                        }
-                    }
-                    // If user chose to recur on the same weekday of every month (e.g. 3rd tuesday)
-                    if (intervalType.equals(MonthlyRecurFragment.EXTRA_VAL_DYNAMIC)) {
-                        int dayOfWeek = recurCal.get(Calendar.DAY_OF_WEEK);
-                        int dayOfWeekInMonth = recurCal.get(Calendar.DAY_OF_WEEK_IN_MONTH);
-
-                        // We will repeat until the other clause (date after last day) is
-                        // reached
-                        if (until) {
-                            numTimes = Integer.MAX_VALUE;
-                        }
-                        else {
-                            Calendar endCal = Calendar.getInstance();
-                            // 2100 as this is when the display functionality breaks
-                            endCal.set(Calendar.YEAR, 2100);
-
-                            endDate = endCal.getTime();
-                        }
-
-                        for (int i = 0; i < numTimes && !recurCal.getTime().after(endDate); i++) {
-                            // Add event to schedule
-                            Event toAdd = new Event(name, ttc, recurCal.getTime());
-
-                            for (int j = mEventSchedule.size(); j <= currIndex; j++) {
-                                mEventSchedule.add(new ArrayList<>());
-                                mDayItemAdapter.notifyItemChanged(j);
-                            }
-
-                            mEventSchedule.get(currIndex).add(toAdd);
-                            mTaskAppViewModel.insert(toAdd);
-                            mDayItemAdapter.notifyItemChanged(currIndex);
-
-                            // Calculate next date
-                            recurCal.add(Calendar.MONTH, interval);
-                            recurCal.set(Calendar.DAY_OF_WEEK, dayOfWeek);
-                            recurCal.set(Calendar.DAY_OF_WEEK_IN_MONTH, dayOfWeekInMonth);
-                            currIndex = getDiff(recurCal.getTime(), mStartDate);
-                        }
-                    }
-                    // If user chose to recur on the same dates of every month (e.g. 2nd and 3rd)
-                    if (intervalType.equals(MonthlyRecurFragment.EXTRA_VAL_SPECIFIC)) {
-                        String[] strs = recur.getString(MonthlyRecurFragment.EXTRA_DAYS)
-                                .split(",");
-                        // Get all the days from the bundle
-                        int[] days = new int[strs.length];
-
-                        for (int i = 0; i < strs.length; i++) {
-                            days[i] = Integer.parseInt(strs[i]);
-                        }
-
-                        Arrays.sort(days);
-
-                        // For each day, add all of it's occurrences to the taskSchedule
-                        for (int day : days) {
-                            // Calculate the first day the task occurs on this day of month
-                            recurCal.setTime(start);
-                            int daysInMonth = recurCal.getActualMaximum(Calendar.DAY_OF_MONTH);
-                            int dayDiff = (day - recurCal.get(Calendar.DAY_OF_MONTH)
-                                    + daysInMonth) % daysInMonth;
-
-                            // Set these fields to help with edge cases
-                            boolean oneLess = (day - recurCal.get(Calendar.DAY_OF_MONTH)) < 0;
-                            boolean offByOne = oneLess && (interval != 1);
-
-                            // Calculate the Date based on the first day he task occurs on
-                            recurCal.add(Calendar.DAY_OF_YEAR, dayDiff);
-                            currIndex = getDiff(recurCal.getTime(), mStartDate);
-
-                            // We will repeat until the other clause (date after last day) is
-                            // reached
-                            if (until) {
-                                numTimes = Integer.MAX_VALUE;
-                            }
-                            else {
-                                Calendar endCal = Calendar.getInstance();
-                                // 2100 as this is when the display functionality breaks
-                                endCal.set(Calendar.YEAR, 2100);
-
-                                endDate = endCal.getTime();
-                            }
-
-                            for (int i = 0; i < numTimes && !recurCal.getTime().after(endDate);
-                                 i++) {
-                                // Add event to schedule
-                                if (!(i == 0 && offByOne) && !(i == numTimes - 1 && oneLess &&
-                                        !offByOne)) {
-                                    Event toAdd = new Event(name, ttc, recurCal.getTime());
-                                    for (int k = mEventSchedule.size(); k <= currIndex; k++) {
-                                        mEventSchedule.add(new ArrayList<>());
-                                        mDayItemAdapter.notifyItemChanged(k);
-                                    }
-                                    mEventSchedule.get(currIndex).add(toAdd);
-                                    mTaskAppViewModel.insert(toAdd);
-                                    mDayItemAdapter.notifyItemChanged(currIndex);
-                                }
-
-                                // Calculate next date
-                                if (i == 0 && offByOne) {
-                                    recurCal.add(Calendar.MONTH, -1);
-                                }
-                                recurCal.add(Calendar.MONTH, interval);
-                                currIndex = getDiff(recurCal.getTime(), mStartDate);
-                            }
-                        }
-                    }
-                }
-                // The user chose to recur yearly
-                else if (YearlyRecurFragment.EXTRA_VAL_TYPE.equals(recurType)) {
-                    // How many years between each recurrence of this event.
-                    int interval = recur.getInt(YearlyRecurFragment.EXTRA_INTERVAL);
-                    // How the event will recur (the 18th, 3rd monday, 18/21st etc.)
-                    String intervalType = recur.getString(YearlyRecurFragment.EXTRA_RECUR_TYPE);
-                    // What months to recur on if necessary.
-                    boolean[] months = new boolean[12];
-
-                    // If necessary, see which months the user chose to recur on
-                    if (intervalType.equals(YearlyRecurFragment.EXTRA_VAL_MULTIPLE_DYNAMIC)
-                            || intervalType.equals(YearlyRecurFragment.EXTRA_VAL_MULTIPLE_STATIC)
-                            || intervalType.equals(YearlyRecurFragment.EXTRA_VAL_SPECIFIC)) {
-                        List<String> monthsStr = Arrays.asList(
-                                recur.getString(YearlyRecurFragment.EXTRA_MONTHS).split(","));
-
-                        // For each month, see if the user chose to include it
-                        months[0] = monthsStr.contains(getString(R.string.jan));
-                        months[1] = monthsStr.contains(getString(R.string.feb));
-                        months[2] = monthsStr.contains(getString(R.string.mar));
-                        months[3] = monthsStr.contains(getString(R.string.apr));
-                        months[4] = monthsStr.contains(getString(R.string.may));
-                        months[5] = monthsStr.contains(getString(R.string.jun));
-                        months[6] = monthsStr.contains(getString(R.string.jul));
-                        months[7] = monthsStr.contains(getString(R.string.aug));
-                        months[8] = monthsStr.contains(getString(R.string.sep));
-                        months[9] = monthsStr.contains(getString(R.string.oct));
-                        months[10] = monthsStr.contains(getString(R.string.nov));
-                        months[11] = monthsStr.contains(getString(R.string.dec));
-                    }
-
-                    // Initialize calendar
-                    Calendar recurCal = Calendar.getInstance();
-                    recurCal.setTime(start);
-
-                    int currIndex = diff;
-
-                    // If user chose to recur on same month/day
-                    if (intervalType.equals(YearlyRecurFragment.EXTRA_VAL_STATIC)) {
-
-                        // Calculate end date
-                        if (!until) {
-                            Calendar endDateCal = Calendar.getInstance();
-                            endDateCal.setTime(start);
-                            endDateCal.add(Calendar.YEAR, interval * (numTimes - 1));
-                            endDate = endDateCal.getTime();
-                        }
-
-                        // Ensure enough space in eventSchedule
-                        for (int i = mEventSchedule.size(); i <= endIndex; i++) {
-                            mEventSchedule.add(new ArrayList<>());
-                            mDayItemAdapter.notifyItemChanged(i);
-                        }
-
-                        while (!recurCal.getTime().after(endDate)) {
-                            // Add event to schedule
-                            Event toAdd = new Event(name, ttc, recurCal.getTime());
-                            for (int k = mEventSchedule.size(); k <= currIndex; k++) {
-                                mEventSchedule.add(new ArrayList<>());
-                                mDayItemAdapter.notifyItemChanged(k);
-                            }
-                            mEventSchedule.get(currIndex).add(toAdd);
-                            mTaskAppViewModel.insert(toAdd);
-                            mDayItemAdapter.notifyItemChanged(currIndex);
-
-                            // Calculate next date
-                            recurCal.add(Calendar.YEAR, interval);
-                            currIndex = getDiff(recurCal.getTime(), mStartDate);
-                        }
-                    }
-                    // If user chose to recur on the same month/weekday (e.g. 3rd Mon of Sep)
-                    if (intervalType.equals(YearlyRecurFragment.EXTRA_VAL_DYNAMIC)) {
-                        int dayOfWeek = recurCal.get(Calendar.DAY_OF_WEEK);
-                        int dayOfWeekInMonth = recurCal.get(Calendar.DAY_OF_WEEK_IN_MONTH);
-
-                        // Set the null fields to their max values to ensure they're ignored by the
-                        // while loop
-                        if (until) {
-                            numTimes = Integer.MAX_VALUE;
-                        }
-                        else {
-                            endDate.setTime(Long.MAX_VALUE);
-                        }
-
-                        for (int i = 0; i < numTimes && !recurCal.getTime().after(endDate); i++) {
-                            // Add event to schedule
-                            Event toAdd = new Event(name, ttc, recurCal.getTime());
-                            for (int k = mEventSchedule.size(); k <= currIndex; k++) {
-                                mEventSchedule.add(new ArrayList<>());
-                                mDayItemAdapter.notifyItemChanged(k);
-                            }
-                            mEventSchedule.get(currIndex).add(toAdd);
-                            mTaskAppViewModel.insert(toAdd);
-                            mDayItemAdapter.notifyItemChanged(currIndex);
-
-                            // Calculate next date
-                            recurCal.add(Calendar.YEAR, interval);
-                            recurCal.set(Calendar.DAY_OF_WEEK, dayOfWeek);
-                            recurCal.set(Calendar.DAY_OF_WEEK_IN_MONTH, dayOfWeekInMonth);
-                            currIndex = getDiff(recurCal.getTime(), mStartDate);
-                        }
-                    }
-                    // If user chose to recur on multiple months on the same day (e.g. Jan/Feb 3rd)
-                    if (intervalType.equals(YearlyRecurFragment.EXTRA_VAL_MULTIPLE_STATIC)) {
-                        // For each month, check if it is recurred on, and if it is, calculate
-                        // events
-                        for (int i = 0; i < months.length; i++) {
-                            if (months[i]) {
-                                // Calculate first year event should be scheduled for
-                                boolean oneLess = (i - recurCal.get(Calendar.MONTH)) < 0;
-                                boolean offByOne = oneLess && (interval != 1);
-                                recurCal = Calendar.getInstance();
-                                recurCal.setTime(start);
-                                recurCal.add(Calendar.MONTH,
-                                        (i - recurCal.get(Calendar.MONTH) + 12) % 12);
-                                currIndex = getDiff(recurCal.getTime(), mStartDate);
-
-                                // Set the null fields to their max values to ensure they're ignored
-                                // by the while loop
-                                if (until) {
-                                    numTimes = Integer.MAX_VALUE;
-                                }
-                                else {
-                                    endDate.setTime(Long.MAX_VALUE);
-                                }
-
-                                for (int j = 0; j < numTimes && !recurCal.getTime()
-                                        .after(endDate); j++) {
-                                    // Add event to schedule
-                                    if (!(j == 0 && offByOne) && !(j == numTimes - 1 && oneLess &&
-                                            !offByOne)) {
-                                        Event toAdd = new Event(name, ttc, recurCal.getTime());
-                                        for (int k = mEventSchedule.size(); k <= currIndex; k++) {
-                                            mEventSchedule.add(new ArrayList<>());
-                                            mDayItemAdapter.notifyItemChanged(k);
-                                        }
-                                        mEventSchedule.get(currIndex).add(toAdd);
-                                        mTaskAppViewModel.insert(toAdd);
-                                        mDayItemAdapter.notifyItemChanged(currIndex);
-                                    }
-
-                                    // Calculate next date
-                                    if (j == 0 && offByOne) {
-                                        recurCal.add(Calendar.YEAR, -1);
-                                    }
-                                    recurCal.add(Calendar.YEAR, interval);
-                                    currIndex = getDiff(recurCal.getTime(), mStartDate);
-                                }
-                            }
-                        }
-                    }
-                    // If the user chose to recur on multiple months on the same weekday (e.g. 3rd
-                    // Monday of Jan/Feb)
-                    if (intervalType.equals(YearlyRecurFragment.EXTRA_VAL_MULTIPLE_DYNAMIC)) {
-                        int dayOfWeek = recurCal.get(Calendar.DAY_OF_WEEK);
-                        int dayOfWeekInMonth = recurCal.get(Calendar.DAY_OF_WEEK_IN_MONTH);
-
-                        // For each month, if user chose month to have event, schedule event
-                        for (int i = 0; i < months.length; i++) {
-                            if (months[i]) {
-                                // Calculate first occurrence of event on this month
-                                boolean oneLess = (i - recurCal.get(Calendar.MONTH)) < 0;
-                                boolean offByOne = oneLess && (interval != 1);
-                                recurCal = Calendar.getInstance();
-                                recurCal.setTime(start);
-                                recurCal.add(Calendar.MONTH,
-                                        (i - recurCal.get(Calendar.MONTH) + 12) % 12);
-                                recurCal.set(Calendar.DAY_OF_WEEK, dayOfWeek);
-                                recurCal.set(Calendar.DAY_OF_WEEK_IN_MONTH, dayOfWeekInMonth);
-                                currIndex = getDiff(recurCal.getTime(), mStartDate);
-
-                                // Set the null fields to their max values to ensure they're ignored
-                                // by the while loop
-                                if (until) {
-                                    numTimes = Integer.MAX_VALUE;
-                                }
-                                else {
-                                    endDate.setTime(Long.MAX_VALUE);
-                                }
-
-                                for (int j = 0; j < numTimes && !recurCal.getTime()
-                                        .after(endDate); j++) {
-                                    // Add event to schedule
-                                    if (!(j == 0 && offByOne) && !(j == numTimes - 1 && oneLess
-                                            && !offByOne)) {
-                                        Event toAdd = new Event(name, ttc, recurCal.getTime());
-                                        for (int k = mEventSchedule.size(); k <= currIndex; k++) {
-                                            mEventSchedule.add(new ArrayList<>());
-                                            mDayItemAdapter.notifyItemChanged(k);
-                                        }
-                                        mEventSchedule.get(currIndex).add(toAdd);
-                                        mTaskAppViewModel.insert(toAdd);
-                                        mDayItemAdapter.notifyItemChanged(currIndex);
-                                    }
-
-                                    // Calculate next date
-                                    if (j == 0 && offByOne) {
-                                        recurCal.add(Calendar.YEAR, -1);
-                                    }
-                                    recurCal.add(Calendar.YEAR, interval);
-                                    recurCal.set(Calendar.DAY_OF_WEEK, dayOfWeek);
-                                    recurCal.set(Calendar.DAY_OF_WEEK_IN_MONTH, dayOfWeekInMonth);
-                                    currIndex = getDiff(recurCal.getTime(), mStartDate);
-                                }
-                            }
-                        }
-                    }
-                    // If user chose to recur on specific months/days (e.g. 2nd/3rd of Jan/Feb)
-                    if (intervalType.equals(YearlyRecurFragment.EXTRA_VAL_SPECIFIC)) {
-                        // Get the days to recur on from user input.
-                        String[] daysStrs = recur.getString(YearlyRecurFragment.EXTRA_DAYS)
-                                .split(",");
-                        int[] days = new int[daysStrs.length];
-
-                        for (int i = 0; i < days.length; i++) {
-                            days[i] = Integer.parseInt(daysStrs[i]);
-                        }
-
-                        // For each month, check if user chose to recur on it, and schedule events
-                        // if they did
-                        for (int i = 0; i < months.length; i++) {
-                            if (months[i]) {
-                                // Calculate earliest possible date to start scheduling events for
-                                // this month
-                                boolean oneLess = (i - recurCal.get(Calendar.MONTH)) < 0;
-                                boolean offByOne = oneLess && (interval != 1);
-                                recurCal = Calendar.getInstance();
-                                recurCal.setTime(start);
-                                recurCal.add(Calendar.MONTH,
-                                        (i - recurCal.get(Calendar.MONTH) + 12) % 12);
-
-                                // Set the null fields to their max values to ensure they're ignored
-                                // by the while loop
-                                if (until) {
-                                    numTimes = Integer.MAX_VALUE;
-                                }
-                                else {
-                                    endDate.setTime(Long.MAX_VALUE);
-                                }
-
-                                // Schedule this month out until the recurrence is set to end.
-                                for (int j = 0; j < numTimes && !recurCal.getTime()
-                                        .after(endDate); j++) {
-                                    for (int day : days) {
-                                        // Make sure to schedule not to schedule days in months
-                                        // they're greater than the max day (e.g. 2/31)
-                                        if (day > recurCal.getActualMaximum
-                                                (Calendar.DAY_OF_MONTH)) {
-                                            continue;
-                                        }
-
-                                        // Set the calendar to this date
-                                        recurCal.set(Calendar.DAY_OF_MONTH, day);
-                                        currIndex = getDiff(recurCal.getTime(), mStartDate);
-
-                                        // Ensure that we don't schedule events for before today or
-                                        // the endDate the user chose
-                                        if (recurCal.getTime().before(start) || recurCal.getTime()
-                                                .after(endDate) || j == 0 && offByOne ||
-                                                j == numTimes - 1 && oneLess && !offByOne) {
-                                            continue;
-                                        }
-
-                                        // Add the event to the data structure.
-                                        Event toAdd = new Event(name, ttc, recurCal.getTime());
-                                        for (int k = mEventSchedule.size(); k <= currIndex; k++) {
-                                            mEventSchedule.add(new ArrayList<>());
-                                            mDayItemAdapter.notifyItemChanged(k);
-                                        }
-                                        mEventSchedule.get(currIndex).add(toAdd);
-                                        mTaskAppViewModel.insert(toAdd);
-                                        mDayItemAdapter.notifyItemChanged(currIndex);
-                                    }
-
-                                    // Handle edge case
-                                    if (j == 0 && offByOne) {
-                                        recurCal.add(Calendar.YEAR, -1);
-                                    }
-                                    recurCal.add(Calendar.YEAR, interval);
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-            // If the user chose to not recur
-            else {
-                // Ensure there is enough space in the eventSchedule
-                for (int i = mEventSchedule.size(); i <= diff; i++) {
-                    mEventSchedule.add(new ArrayList<>());
-                    mDayItemAdapter.notifyItemChanged(i);
-                }
-
-                // Schedule the event
-                Event toAdd = new Event(name, ttc, start);
-                mEventSchedule.get(diff).add(toAdd);
-                mTaskAppViewModel.insert(toAdd);
-                mDayItemAdapter.notifyItemChanged(diff);
-            }
+        List<Integer> updatedIndices = mLogicSubsystem.addItem(data);
+
+        if (updatedIndices == null) {
+            Toast.makeText(this, "Error occurred when adding new item, try again.",
+                    Toast.LENGTH_LONG).show();
+            return;
         }
-        // If the item type is Task
-        else if (type.equals(AddItem.EXTRA_VAL_TASK)) {
-            // Get the fields from the Bundle
-            String name = result.getString(AddItem.EXTRA_NAME);
-            int timeToComplete = Integer.parseInt(result.getString(AddItem.EXTRA_TTC));
-            String ecd = result.getString(AddItem.EXTRA_ECD);
-            String dd = result.getString(AddItem.EXTRA_DUE);
-            String parents = result.getString(AddItem.EXTRA_PARENTS);
 
-            // Convert the earliest completion date String to a MyTime
-            Date early;
-            try {
-                early = Task.dateFormat.parse(ecd);
-            }
-            catch (Exception e) {
-                Toast.makeText(this, e.getMessage(), Toast.LENGTH_SHORT).show();
-                return;
-            }
-
-            // Convert the due date String to a MyTime
-            Date due;
-            try {
-                due = Task.dateFormat.parse(dd);
-            }
-            catch (Exception e) {
-                Toast.makeText(this, e.getMessage(), Toast.LENGTH_SHORT).show();
-                return;
-            }
-
-            Task toAdd = new Task(name, early, due, timeToComplete);
-
-            // The parents string in the Bundle is a String of the format "n1,n2,n3,...nN,"
-            // where each nx is an index to a Task in tasks that should be used as a parent
-            // for the task to be added.
-            String[] parentIndices = parents.split(",");
-            for (String parentIndex : parentIndices) {
-                if (!parentIndex.equals("-1")) {
-                    Task parent = mTasks.get(Integer.parseInt(parentIndex));
-                    toAdd.addParent(parent);
-                    parent.addChild(toAdd);
-                }
-            }
-
-            // Add the event to the data structure and database
-            mTasks.add(toAdd);
-            mTaskAppViewModel.insert(toAdd);
+        for (int index : updatedIndices) {
+            mDayItemAdapter.notifyItemChanged(index);
         }
 
         // As the task dependency graph has been updated, we must reoptimize it
-        Optimize(true);
+        Optimize();
         // Show recycler as Optimize is finished
         mVF.setDisplayedChild(1);
     }
@@ -832,78 +108,38 @@ public class MainActivity extends AppCompatActivity implements ClickListener {
     /**
      * Calls the Optimizer to find an optimal schedule for the user's tasks, given the user's
      * scheduled events.
-     *
-     * @param refresh true refreshes recycler, false doesn't allows us to call Optimize() before
-     *                recycler initialization
      */
-    private void Optimize(boolean refresh) {
-        Optimizer opt = new Optimizer();
-        ArrayList<Task> changedTasks = opt.Optimize
-                (mTasks, mTaskSchedule, mEventSchedule, mStartDate, mTodayTime);
-
-        // Delete any empty lists at the end of the taskSchedule.
-        int taskLowIndex = mTaskSchedule.size();
-        for (int i = mTaskSchedule.size() - 1; i >= 0; i--) {
-            if (mTaskSchedule.get(i).size() != 0) {
-                break;
-            }
-            else {
-                taskLowIndex--;
-            }
-        }
-        if (mTaskSchedule.size() > taskLowIndex) {
-            mTaskSchedule.subList(taskLowIndex, mTaskSchedule.size()).clear();
-        }
-
-        // Delete any empty lists at the end of the event schedule.
-        int eventLowIndex = mEventSchedule.size();
-        for (int i = mEventSchedule.size() - 1; i >= 0; i--) {
-            if (mEventSchedule.get(i).size() != 0) {
-                break;
-            }
-            else {
-                eventLowIndex--;
-            }
-        }
-        int eventScheduleSize = mEventSchedule.size();
-        if (eventScheduleSize > eventLowIndex) {
-            mEventSchedule.subList(eventLowIndex, eventScheduleSize).clear();
-
-            // Remove hanging days from adapter as well
-            eventLowIndex = Math.max(eventLowIndex, taskLowIndex);
-            if (eventScheduleSize > eventLowIndex) {
-                mDayItemAdapter.notifyItemRangeRemoved(eventLowIndex, eventScheduleSize);
-            }
-        }
+    private void Optimize() {
+        Pair<Pair<Integer, Integer>, List<Pair<Integer, Integer>>> updatedInfo =
+                mLogicSubsystem.Optimize();
 
         // As the Optimizer may have changed tasks' dates, we must refresh the recycler
-        if (refresh) {
-            mDayItemAdapter.mDayItemList = DayItemList();
+        int eventLowIndex = updatedInfo.getFirst().getFirst();
+        int eventScheduleSize = updatedInfo.getFirst().getSecond();
+        List<Pair<Integer, Integer>> changedIndices = updatedInfo.getSecond();
 
-            // Tell the recycler about moved tasks.
-            for (Task t : changedTasks) {
-                int oldIndex = getDiff(t.getDoDate(), mStartDate);
-                int newIndex = getDiff(t.getWorkingDoDate(), mStartDate);
-
-                // We must use changed instead of moved, as an item in the dayItem entry is moved
-                // not the dayItem itself.
-                if (oldIndex >= 0) {
-                    mDayItemAdapter.notifyItemChanged(oldIndex);
-                }
-                if (newIndex >= mDayItemAdapter.getItemCount()) {
-                    mDayItemAdapter.notifyItemInserted(newIndex);
-                }
-                else {
-                    mDayItemAdapter.notifyItemChanged(newIndex);
-                }
-            }
+        if (eventScheduleSize > eventLowIndex) {
+            mDayItemAdapter.notifyItemRangeRemoved(eventLowIndex, eventScheduleSize);
         }
 
-        // Update the task with the new do date, and reflect this change in the database.
-        for (Task t : changedTasks) {
-            t.setDoDate(t.getWorkingDoDate());
-            Log.d("TEST", String.valueOf(t.getID()));
-            mTaskAppViewModel.update(t);
+        mDayItemAdapter.mDayItemList = mLogicSubsystem.DayItemList();
+
+        // Tell the recycler about moved tasks.
+        for (Pair<Integer, Integer> indices : changedIndices) {
+            int oldIndex = indices.getFirst();
+            int newIndex = indices.getSecond();
+
+            // We must use changed instead of moved, as an item in the dayItem entry is moved
+            // not the dayItem itself.
+            if (oldIndex >= 0) {
+                mDayItemAdapter.notifyItemChanged(oldIndex);
+            }
+            if (newIndex >= mDayItemAdapter.getItemCount()) {
+                mDayItemAdapter.notifyItemInserted(newIndex);
+            }
+            else {
+                mDayItemAdapter.notifyItemChanged(newIndex);
+            }
         }
     }
 
@@ -923,96 +159,21 @@ public class MainActivity extends AppCompatActivity implements ClickListener {
         // startDate is our representation for the current date upon the launch of TaskApp.
         mStartDate = clearDate(new Date());
 
-        mTodayTime = 0;
-
         // Get todayTime from shared preferences
         SharedPreferences sp = getSharedPreferences(PREF_FILE, MODE_PRIVATE);
         Date todayTimeDate = new Date(sp.getLong(PREF_DAY, -1L));
-
+        int todayTime = 0;
         if (!todayTimeDate.before(mStartDate)) {
-            mTodayTime = sp.getInt(PREF_TIME, 0);
+            todayTime = sp.getInt(PREF_TIME, 0);
         }
 
-        ViewModelStoreOwner vmso = this;
+        mLogicSubsystem = new LogicSubsystem(this, todayTime);
 
-        // Populate from database;
-        Thread thread = new Thread() {
-            public void run() {
-                mTaskAppViewModel = new ViewModelProvider(vmso).get(TaskAppViewModel.class);
-            }
-        };
-        thread.start();
-        try {
-            thread.join();
-        } catch (InterruptedException e) {
-            e.printStackTrace();
-        }
-
-        // Get tasks from database
-        mTasks = mTaskAppViewModel.getAllTasks();
-        mTasks = (mTasks == null) ? new ArrayList<>() : mTasks;
-
-        ArrayList<Task> overdueTasks = new ArrayList<>(); // Tasks that are overdue.
-
-        // Add tasks to taskSchedule/add parents
-        for (Task t : mTasks) {
-            // Calculate how many days past today's date this task is scheduled for. Used to
-            // index into taskSchedule
-            Date doDate = t.getDoDate();
-            int index = getDiff(doDate, mStartDate);
-
-            if (t.getEarlyDate().before(mStartDate)) {
-                t.setEarlyDate(mStartDate);
-                mTaskAppViewModel.update(t);
-            }
-
-            // Adds file to taskSchedule if it is scheduled for today or later.
-            if (index >= 0) {
-                // Make sure taskSchedule is big enough
-                for (int i = mTaskSchedule.size(); i <= index; i++) {
-                    mTaskSchedule.add(new ArrayList<>());
-                }
-
-                // Add to taskSchedule
-                mTaskSchedule.get(index).add(t);
-            } else {
-                overdueTasks.add(t);
-            }
-        }
-
-        // Add parent/child structure to task lists, as Room DB cannot do this
-        for (Task t : mTasks) {
-            if (t.getParentArr().size() != 0) {
-                for (Task other : mTasks) {
-                    if (t.getParentArr().contains(other.getID())) {
-                        t.addParent(other);
-                        other.addChild(t);
-                    }
-                }
-            }
-        }
+        String[] overdueNames = mLogicSubsystem.getOverdueTasks();
 
         // Prompt the user with a dialog containing overdue tasks so they can mark overdue tasks
         // so taskapp can reoptimize the schedule if some tasks are overdue.
-        if (overdueTasks.size() != 0) {
-
-            String[] overdueNames = new String[overdueTasks.size()];
-
-            // Create a list of overdue task names for the dialog
-            for (int i = 0; i < overdueNames.length; i++) {
-                Task t = overdueTasks.get(i);
-
-                if (t.getDoDate().getTime() == 0) {
-                    mTasks.remove(t);
-                    mTaskAppViewModel.delete(t);
-                    continue;
-                }
-
-                Date tDate = t.getDueDate();
-                overdueNames[i] = String.format(getString(R.string.due_when), t.getName(),
-                        Task.dateFormat.format(tDate));
-            }
-
+        if (overdueNames.length != 0) {
             // List of indices to tasks that were completed.
             ArrayList<Integer> selectedItems = new ArrayList<>();
 
@@ -1056,28 +217,7 @@ public class MainActivity extends AppCompatActivity implements ClickListener {
                                  */
                                 @Override
                                 public void onClick(DialogInterface di, int index) {
-                                    // As the user has marked these tasks as completed, remove them.
-                                    // Pass in false as the user completed them on a prior day.
-                                    for (int i = 0; i < selectedItems.size(); i++) {
-                                        Complete(overdueTasks.get(selectedItems.get(i)),
-                                                false, false);
-                                    }
-
-                                    // Change due date for overdue tasks if it has already been
-                                    // passed to today.
-                                    if (overdueTasks.size() != selectedItems.size()) {
-                                        for (int i = 0; i < overdueTasks.size(); i++) {
-                                            if (selectedItems.contains(i)) {
-                                                continue;
-                                            }
-
-                                            Task t = overdueTasks.get(i);
-
-                                            if (t.getDueDate().before(mStartDate)) {
-                                                t.setDueDate(mStartDate);
-                                            }
-                                        }
-                                    }
+                                    mLogicSubsystem.updateOverdueTasks(selectedItems);
 
                                     finishProcessing(true);
                                 }
@@ -1093,44 +233,19 @@ public class MainActivity extends AppCompatActivity implements ClickListener {
     }
 
     /**
-     * Extremely sketchy, but allows us to wait to finish processing events and displaying the
-     * recycler until the user has completed the overdue tasks dialog
+     * Now that the user has handled overdue tasks, we're ready to start the app. This function
+     * adds data to the recycler and moves to the working fragment.
      *
      * @param reoptimize If the task list has changed due to overdue tasks, we might have to
      *                   reoptimize it
      */
     private void finishProcessing(boolean reoptimize) {
-        // Get the event list.
-        List<Event> events = mTaskAppViewModel.getAllEvents();
-
-        events = (events == null) ? new ArrayList<>() : events;
-
-        // Add events from the database into the eventSchedule
-        for (Event e : events) {
-            // Calculate how many days past today's date this event is scheduled for. Used to
-            // index into eventSchedule
-            Date doDate = e.getDoDate();
-            int doDateIndex = getDiff(doDate, mStartDate);
-
-            // Add the events to the list if they aren't for an earlier date
-            if (!doDate.before(mStartDate)) {
-                for (int j = mEventSchedule.size(); j <= doDateIndex; j++) {
-                    mEventSchedule.add(new ArrayList<>());
-                }
-
-                mEventSchedule.get(doDateIndex).add(e);
-            }
-        }
-
-        // If tasks were changed, make sure to reoptimize the schedule in case it's necessary
-        if (reoptimize) {
-            Optimize(false);
-        }
+        List<DayItem> recyclerData = mLogicSubsystem.prepForDisplay(reoptimize);
 
         // Initialize the main recyclerview with data calculated in helper function DayItemList
         RecyclerView dayRecyclerView = findViewById(R.id.main_recyclerview);
         LinearLayoutManager layoutManager = new LinearLayoutManager(MainActivity.this);
-        mDayItemAdapter = new DayItemAdapter(DayItemList(), this);
+        mDayItemAdapter = new DayItemAdapter(recyclerData, this);
         dayRecyclerView.setAdapter(mDayItemAdapter);
         dayRecyclerView.setLayoutManager(layoutManager);
 
@@ -1143,7 +258,7 @@ public class MainActivity extends AppCompatActivity implements ClickListener {
         // Make visible the main content
         mVF = findViewById(R.id.vf);
 
-        if (mTaskSchedule.size() != 0 || mEventSchedule.size() != 0) {
+        if (!mLogicSubsystem.isEmpty()) {
             mVF.setDisplayedChild(1);
         }
         else {
@@ -1168,7 +283,7 @@ public class MainActivity extends AppCompatActivity implements ClickListener {
         SharedPreferences sp = getSharedPreferences(PREF_FILE, MODE_PRIVATE);
         SharedPreferences.Editor edit = sp.edit();
         edit.putLong(PREF_DAY, mStartDate.getTime());
-        edit.putInt(PREF_TIME, mTodayTime);
+        edit.putInt(PREF_TIME, mLogicSubsystem.getTodayTime());
         edit.apply();
     }
 
@@ -1178,159 +293,11 @@ public class MainActivity extends AppCompatActivity implements ClickListener {
     private void intentAddItem() {
         Intent intent = new Intent(this, AddItem.class);
 
-        // Create an ArrayList of task names so the multiselect dialog can use it to choose parent
-        // tasks
-        ArrayList<String> taskNames = new ArrayList<>();
-        for (Task t : mTasks) {
-            Date tDate = t.getDueDate();
-            taskNames.add(String.format(getString(R.string.due_when), t.getName(),
-                    Task.dateFormat.format(tDate)));
-        }
+        // Get a list of task names for prerequisite list
+        ArrayList<String> taskNames = mLogicSubsystem.getTaskNames();
 
         intent.putStringArrayListExtra(EXTRA_TASKS, taskNames);
         mStartForResult.launch(intent);
-    }
-
-    /**
-     * Get the EventItem and TaskItem lists for the day i days past today's date. Is used as a
-     * helper function so when the recycler needs to be refreshed, it only needs to get the DayItem
-     * for the updated days, not the entire schedule.
-     *
-     * @param i How many days past today's date to get the schedule for
-     *
-     * @return A DayItem representing that day's schedule
-     */
-    private DayItem DayItemHelper(int i) {
-        // Fields for the DayItem
-        String dayString;
-        List<EventItem> events;
-        List<TaskItem> tasks;
-
-        // MyTime representing the date i days past today's date
-        Calendar currCal = Calendar.getInstance();
-        currCal.add(Calendar.DAY_OF_YEAR, i);
-        Date curr = clearDate(currCal.getTime());
-
-        // Number representing totalTime this date has scheduled. If it's today's date, add
-        // todayTime to represent the time already completed tasks took.
-        int totalTime = (i == 0) ? mTodayTime : 0;
-
-        // Adds the total event time for the day to the total time
-        if (i < mEventSchedule.size() && mEventSchedule.get(i).size() > 0) {
-            for (int j = 0; j < mEventSchedule.get(i).size(); j++) {
-                Event event = mEventSchedule.get(i).get(j);
-                totalTime += event.getLength();
-            }
-        }
-
-        // Adds the total task time for the day to the total time
-        if (i < mTaskSchedule.size() && mTaskSchedule.get(i).size() > 0) {
-            for (int j = 0; j < mTaskSchedule.get(i).size(); j++) {
-                Task task = mTaskSchedule.get(i).get(j);
-
-                totalTime += task.getTimeToComplete();
-            }
-        }
-
-        // Set the fields
-        dayString = String.format(getString(R.string.schedule_for), Task.dateFormat.format(curr),
-                totalTime);
-        events = EventItemList(i);
-        tasks = TaskItemList(i);
-
-        return new DayItem(dayString, events, tasks, i);
-    }
-
-    /**
-     * Builds a DayItem List representation of a user's tasks/events
-     *
-     * @return a DayItem List representation of a user's tasks/events
-     */
-    private List<DayItem> DayItemList() {
-        // The list of DayItem's to be displayed in the recycler
-        List<DayItem> itemList = new ArrayList<>();
-
-        // Generate a DayItem for the date i days past today's date
-        for (int i = 0; i < mTaskSchedule.size() || i < mEventSchedule.size(); i++) {
-            itemList.add(DayItemHelper(i));
-        }
-
-        return itemList;
-    }
-
-    /**
-     * Builds an EventItem List representation of a user's events on a given day
-     *
-     * @param index The index into the data structure representing the day
-     *
-     * @return an EventItem List representation of a user's events on a given day
-     */
-    private List<EventItem> EventItemList(int index) {
-        // The list of EventItems representing the events scheduled for index days past today's date
-        List<EventItem> itemList = new ArrayList<>();
-
-        // Add all the events scheduled for the given day to itemList, if any are scheduled
-        if (index < mEventSchedule.size() && mEventSchedule.get(index).size() > 0) {
-            for (int j = 0; j < mEventSchedule.get(index).size(); j++) {
-                // Fields for itemList
-                String name;
-                String timespan;
-
-                // Get the jth event from the given date
-                Event event = mEventSchedule.get(index).get(j);
-
-                // Get the start/end time in MyTime objects
-                Date eventTime = event.getDoDate();
-                Calendar endCal = Calendar.getInstance();
-                endCal.setTime(eventTime);
-                endCal.add(Calendar.MINUTE, event.getLength());
-                Date endTime = clearTime(endCal.getTime());
-
-                // Format the event name as Name: StartTime-EndTime
-                name = event.getName();
-                timespan = Event.timeFormat.format(eventTime) + "-" +
-                        Event.timeFormat.format(endTime);
-
-                itemList.add(new EventItem(name, timespan, j));
-            }
-        }
-        return itemList;
-    }
-
-    /**
-     * Builds a TaskItem List representation of a user's tasks on a given day
-     *
-     * @param index The index into the data structure representing the day
-     *
-     * @return a TaskItem List representation of a user's tasks on a given day
-     */
-    private List<TaskItem> TaskItemList(int index) {
-        // The list of TaskItems representing the tasks scheduled for the date index days past
-        // today's date
-        List<TaskItem> itemList = new ArrayList<>();
-
-        // Add all the tasks scheduled for the given date to itemList
-        if (index < mTaskSchedule.size() && mTaskSchedule.get(index).size() > 0) {
-            for (int j = 0; j < mTaskSchedule.get(index).size(); j++) {
-                // DayItem's only field
-                String name;
-                boolean completable;
-
-                // Get the jth task scheduled for the given day.
-                Task task = mTaskSchedule.get(index).get(j);
-
-                // Create the name in the format NAME (TTC minutes to complete)
-                name = task.getName() + "\n" + String.format(getString(R.string.minutes_to_complete),
-                        task.getTimeToComplete());
-
-                completable = (task.getEarlyDate().equals(mStartDate))
-                        && (task.getParents().size() == 0);
-
-                itemList.add(new TaskItem(name, j, completable));
-            }
-        }
-
-        return itemList;
     }
 
     /**
@@ -1380,39 +347,64 @@ public class MainActivity extends AppCompatActivity implements ClickListener {
         // Remove the given task from the task dependency graph
         mVF.setDisplayedChild(0);
 
-        if (action == 0) {
-            if (day == -1 || mTaskSchedule.get(day).size() <= position) {
-                mVF.setDisplayedChild(1);
-                return;
-            }
+        boolean success = mLogicSubsystem.onButtonClick(position, day, action);
 
-            mTaskAppViewModel.delete(mTaskSchedule.get(day).get(position));
-            Complete(mTaskSchedule.get(day).get(position), true, true);
-        }
-        // Remove the given task from the task dependency graph without completion time dialog
-        if (action == 1) {
-            if (day == -1 || mTaskSchedule.get(day).size() <= position) {
+        if (action == 0 || action == 1) {
+            if (!success) {
                 mVF.setDisplayedChild(1);
                 return;
             }
-            mTaskAppViewModel.delete(mTaskSchedule.get(day).get(position));
-            Complete(mTaskSchedule.get(day).get(position), false, true);
+        }
+
+        if (action == 0) {
+            int completionTime = -1;
+
+            // Prompt the user to ask how long it took to complete the task, and add this time to
+            // todayTime to prevent the user from being overscheduled on today's date.
+            AlertDialog.Builder builder = new AlertDialog.Builder(this);
+            builder.setMessage(getString(R.string.complete_dialog_message));
+
+            builder.setTitle(R.string.complete_dialog_title);
+
+            final EditText input = new EditText(this);
+            input.setInputType(InputType.TYPE_CLASS_NUMBER);
+            builder.setView(input);
+
+            builder.setPositiveButton(R.string.complete_task, (dialogInterface, i) -> {
+                mLogicSubsystem.addTodayTime(Integer.parseInt(input.getText().toString()));
+
+                // As the task dependency graph has been updated, we must reoptimize it
+                Optimize();
+
+                // If there are any tasks/events scheduled, show the recycler
+                if (!mLogicSubsystem.isEmpty()) {
+                    mVF.setDisplayedChild(1);
+                }
+                // If there aren't any, show the "add a task" fragment
+                else {
+                    mVF.setDisplayedChild(2);
+                }
+            });
+
+            builder.show();
+            return;
+        }
+        else if (action == 1) {
+            Optimize();
         }
         // Remove the given event from the schedule and re-optimize.
-        if (action == 2) {
-            if (day == -1 || mEventSchedule.get(day).size() <= position) {
+        else if (action == 2) {
+            if (!success) {
                 mVF.setDisplayedChild(1);
                 return;
             }
-            mTaskAppViewModel.delete(mEventSchedule.get(day).get(position));
-            mEventSchedule.get(day).remove(position);
-            mDayItemAdapter.mDayItemList.set(day, DayItemHelper(day));
+            mDayItemAdapter.mDayItemList.set(day, mLogicSubsystem.DayItemHelper(day));
             mDayItemAdapter.notifyItemChanged(day);
-            Optimize(true);
         }
 
+
         // If there are any tasks/events scheduled, show the recycler
-        if (mTaskSchedule.size() != 0 || mEventSchedule.size() != 0) {
+        if (!mLogicSubsystem.isEmpty()) {
             mVF.setDisplayedChild(1);
         }
         // If there aren't any, show the "add a task" fragment
@@ -1423,7 +415,6 @@ public class MainActivity extends AppCompatActivity implements ClickListener {
         // If the FAB is currently hidden, show the FAB again, to prevent it from being lost as the
         // FAB hides if you scroll down currently, and if we don't do this and the recycler doesn't
         // have enough content to scroll, the FAB will be lost until a restart.
-        //noinspection unchecked
         HideBottomViewOnScrollBehavior<FloatingActionButton> fabBehavior =
                 (HideBottomViewOnScrollBehavior<FloatingActionButton>)
                         ((CoordinatorLayout.LayoutParams) mBinding.fab.getLayoutParams())
